+++ gcdprotogen/types/chrome_profiler_types.go
@@ -12,7 +12,7 @@ type ChromeProfilerCPUProfileNode struct {
 	LineNumber int `json:"lineNumber"` // 1-based line number of the function start position.
 	ColumnNumber int `json:"columnNumber"` // 1-based column number of the function start position.
 	HitCount int `json:"hitCount"` // Number of samples where this node was on top of the call stack.
-	CallUID float `json:"callUID"` // Call UID.
+	CallUID float64 `json:"callUID"` // Call UID.
 	Children []*ChromeProfilerCPUProfileNode `json:"children"` // Child nodes.
 	DeoptReason string `json:"deoptReason"` // The reason of being not optimized. The function may be deoptimized or marked as don't optimize.
 	Id int `json:"id"` // Unique id of the node.
@@ -22,9 +22,9 @@ type ChromeProfilerCPUProfileNode struct {
 // Profile.
 type ChromeProfilerCPUProfile struct {
 	Head *ChromeProfilerCPUProfileNode `json:"head"` 
-	StartTime float `json:"startTime"` // Profiling start time in seconds.
-	EndTime float `json:"endTime"` // Profiling end time in seconds.
+	StartTime float64 `json:"startTime"` // Profiling start time in seconds.
+	EndTime float64 `json:"endTime"` // Profiling end time in seconds.
 	Samples []int `json:"samples,omitempty"` // Ids of samples top nodes.
-	Timestamps []float `json:"timestamps,omitempty"` // Timestamps of the samples in microseconds.
+	Timestamps []float64 `json:"timestamps,omitempty"` // Timestamps of the samples in microseconds.
 }
